pyproject.toml
[project]
name = "grit-thesaurus-query-expansion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
]
